ai_tools_suite_streamlit/requirements.txt
streamlit
openai
pandas
matplotlib
plotly
PyMuPDF
python-dotenv